+++ dcm_import_module/handlers.py
@@ -3,8 +3,8 @@
 from typing import Mapping
 from pathlib import Path
 
-from dcm_common.services import handlers, TargetPath, UUID
-from data_plumber_http import Property, Object, String, Url, Boolean
+from dcm_common.services import handlers, UUID
+from data_plumber_http import Property, Object, String, Url, Boolean, FileSystemObject
 
 from dcm_import_module.plugins import IEImportPlugin
 from dcm_import_module.models.import_config import Target, ImportConfigIPs
@@ -47,49 +47,44 @@ def get_ies_import_handler(acceptable_plugins: Mapping[str, IEImportPlugin]):
     ).assemble()
 
 
-def get_ips_import_handler(cwd: Path):
-    """
-    Returns parameterized handler (based on cwd)
-    """
-    return Object(
-        properties={
-            Property("import", name="import_", required=True): Object(
-                model=ImportConfigIPs,
-                properties={
-                    Property("target", required=True): Object(
-                        model=Target,
-                        properties={
-                            Property("path", required=True): TargetPath(
-                                _relative_to=cwd, cwd=cwd, is_dir=True
-                            )
-                        },
-                        accept_only=["path"],
-                    ),
-                    Property("batch", default=True): Boolean(),
-                    Property("test", default=False): Boolean(),
-                },
-                accept_only=[
-                    "target",
-                    "batch",
-                    "test",
-                ],
-            ),
-            Property(
-                "specificationValidation", name="spec_validation"
-            ): Object(free_form=True),
-            Property("objectValidation", name="obj_validation"): Object(
-                free_form=True
-            ),
-            Property("token"): UUID(),
-            Property("callbackUrl", name="callback_url"): Url(
-                schemes=["http", "https"]
-            ),
-        },
-        accept_only=[
-            "import",
-            "specificationValidation",
-            "objectValidation",
-            "token",
-            "callbackUrl",
-        ],
-    ).assemble()
+ips_import_handler = Object(
+    properties={
+        Property("import", name="import_", required=True): Object(
+            model=ImportConfigIPs,
+            properties={
+                Property("target", required=True): Object(
+                    model=Target,
+                    properties={
+                        Property("path", required=True): FileSystemObject(),
+                        Property("hotfolderId", "hotfolder_id"): String(),
+                    },
+                    accept_only=["path", "hotfolderId"],
+                ),
+                Property("batch", default=True): Boolean(),
+                Property("test", default=False): Boolean(),
+            },
+            accept_only=[
+                "target",
+                "batch",
+                "test",
+            ],
+        ),
+        Property("specificationValidation", name="spec_validation"): Object(
+            free_form=True
+        ),
+        Property("objectValidation", name="obj_validation"): Object(
+            free_form=True
+        ),
+        Property("token"): UUID(),
+        Property("callbackUrl", name="callback_url"): Url(
+            schemes=["http", "https"]
+        ),
+    },
+    accept_only=[
+        "import",
+        "specificationValidation",
+        "objectValidation",
+        "token",
+        "callbackUrl",
+    ],
+).assemble()