+++ dcm_import_module/models/__init__.py
@@ -3,6 +3,7 @@ from .import_result import ImportResult
 from .ie import IE
 from .ip import IP
 from .import_config import ImportConfigIEs, ImportConfigIPs
+from .hotfolder import Hotfolder
 
 
 __all__ = [
@@ -12,4 +13,5 @@ __all__ = [
     "ImportResult",
     "ImportConfigIPs",
     "ImportConfigIEs",
+    "Hotfolder",
 ]